backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
django-sesame==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.4
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
websockets==12.0
